pyproject.toml
[project]
name = "inatinq"
version = "0.0.0"
description = "The iNaturalist-INQUIRE project to add INQUIRE based text semantic search to the iNaturalist platform."
readme = "README.md"
authors = [
    { name = "Varun Agrawal", email = "varunagrawal@gatech.edu" },
    { name = "Ketan Bhardwaj", email= "ketanbj@cc.gatech.edu" },
    { name = "Varun Narayan", email= "varun.narayan@gatech.edu" },
]
requires-python = ">= 3.11"
dependencies = [
    "inquire_api",
    "inatinqperf",
    "inat-toolkit",
    "datasets>=4.4.1",
    "docker>=7.1.0",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "ipython>=8.37.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-loguru>=0.4.0",
    "ruff>=0.13.0",
]

[tool.uv.sources]
inquire_api = { workspace = true }
inatinqperf = { workspace = true }
inat_toolkit = { workspace = true }

[tool.uv.workspace]
members = [
    "inquire_api",
    "iNatInqPerf",
    "inat_toolkit",
]

[project.license]
text = "MIT"

[project.urls]
Homepage = "https://github.com/gt-sse-center/inatinq"
Documentation = "https://github.com/gt-sse-center/inatinq"
Repository = "https://github.com/gt-sse-center/inatinq"
